dora-dm-tac/pyproject.toml
[project]
name = "dora-dm-tac"
version = "0.3.12"
authors = [{ name = "Xu Runtian", email = "xuruntian03@163.com" }]
description = "Dora Node for capturing information from dm-tac"
license = "MIT"
readme = "README.md"
requires-python = ">=3.9, <3.11"

dependencies = [
    "dora-rs",
    "dora-rs-cli",
    "dora-rs >= 0.3.9",
    "numpy==1.24.4",
    "opencv-python==4.10.0.84",
    "opencv-contrib-python==4.10.0.84",
    "scipy==1.10.1",
    "setuptools==45.2.0",
    "cupy-cuda12x==12.3.0", # depending on your cuda version
    'pyudev; platform_system=="Linux"',  # For Linux platform

]
[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]
[project.scripts]
dora-dm-tac = "dora_dm_tac_.main:main"

[project.optional-dependencies]
# dev dependencies
dev = ["ruff", "pre-commit"]
# test dependencies
test = ["pytest"]
# doc dependencies
docs = ["sphinx"]

# ruff format and lint config
[tool.ruff]
line-length = 100
indent-width = 2
# ruff format igonres
exclude = [
    ".git", ".venv", "__pypackages__",
    "build", "dist", "node_modules"
]

[tool.ruff.lint]
# 默认启用的规则集
select = ["E4", "E7", "E9", "F"]  # 基础pycodestyle和Pyflakes规则
ignore = []
fixable = ["ALL"]  # 所有规则都可自动修复
unfixable = []
extend-select = [
  "UP",   # Ruff's UP rule
  "PERF", # Ruff's PERF rule
  "RET",  # Ruff's RET rule
  "RSE",  # Ruff's RSE rule
  "NPY",  # Ruff's NPY rule
  "N",    # Ruff's N rule
  "I",   # isort (import sorting)
  "ANN"
]

# 允许下划线前缀的未使用变量
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# 格式化风格（与Black兼容）
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true# 默认不格式化文档字符串中的代码示例
# setuptools config
[tool.setuptools.packages.find]
where = ["src"]  # 源码目录
include = [
  "dora_dm_tac",
  "DM-Tac-WM/dmrobotics"
  # Add other dirs
  # "other_dir",
]

# add other configs
